requirements.txt
Flask
twilio
google-generativeai
python-decouple
sqlalchemy
psycopg2-binary
python-multipart